requirements.txt
flask==3.0.0
gunicorn==21.2.0
jpype1==1.5.0